final changes before first prerelease
