+++ src/charts/index.ts
@@ -1,8 +1,8 @@
 export {type BarChart, generateBarchartConfigs} from "./barchart";
-export type {Chart, ChartType} from "./common";
+export type {BaseChart, ChartType} from "./common";
 export type {AxisSeries, CategoryAxis, Datagroup, LevelCaption} from "./datagroup";
 export {type DonutChart, generateDonutConfigs} from "./donut";
-export {generateCharts} from "./generator";
+export {type Chart, generateCharts } from "./generator";
 export {type ChoroplethMap, generateChoroplethMapConfigs} from "./geomap";
 export {type LinePlot, generateLineplotConfigs} from "./lineplot";
 export {type StackedArea, generateStackedareaConfigs} from "./stackedarea";
