+++ src/index.ts
@@ -1,3 +1,3 @@
 export {Column, LevelColumn, MeasureColumn, PropertyColumn} from "./toolbox/columns";
-export {Chart, ChartType, generateCharts} from "./toolbox/generateCharts";
+export {Chart, generateCharts, ChartType} from "./charts";
 export {ChartLimits, Dataset} from "./types";
