+++ src/toolbox/chartConfigs.ts
@@ -1,10 +1,10 @@
 import {assign} from "d3plus-common";
 import {includes} from "lodash-es";
+import type {ChartType} from "../charts/common";
 import type {D3plusConfig} from "../d3plus";
 import type {UIParams} from "../types";
 import type {Chart} from "./charts";
 import {getCaption} from "./columns";
-import type {ChartType} from "./generateCharts";
 import {propertySorterFactory} from "./sort";
 import {chartTitleGenerator} from "./title";
 import {tooltipGenerator} from "./tooltip";
