+++ tsup.config.js
@@ -3,7 +3,10 @@ import {defineConfig} from "tsup";
 
 export default defineConfig(options => ({
   clean: !options.watch,
-  entry: ["src/charts/index.ts", "src/react/index.ts"],
+  entry: {
+    index: "src/charts/index.ts",
+    react: "src/react/index.ts"
+  },
   format: ["cjs", "esm"],
   outExtension({format}) {
     return {js: `.${format}.js`};
